requirements.txt

future
bs4
telegraph
emoji
lxml
gitpython==3.1.11
heroku3==4.2.3
beautifulsoup4
requests
sqlalchemy
python-telegram-bot==12.8
psycopg2-binary
feedparser
pynewtonmath
spongemock
zalgo-text
geopy
nltk
psutil
aiohttp>=2.2.5
Pillow>=4.2.0
CurrencyConverter
googletrans==3.1.0a0
jikanpy
speedtest-cli
coffeehouse
regex
bleach
git+https://github.com/starry69/python-markdown2.git
wikipedia
telethon==1.20
spamwatch
alphabet_detector
pyrate-limiter
cachetools
pyrogram>=1.1.13
tgCrypto>=1.2.2
fontTools
gtts
tswift
pytz
youtube_dl
PyYAML>=5.3.1
redis
html2text
pythonping
html_telegraph_poster
pytube
youtube-search-python
instantmusic
PyLyrics
hachoir
pybase64
pySmartDL
validators
bing_image_downloader
google-api-python-client
